# console/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorizationRequest": {
        "additionalProperties": false,
        "description": "A charge asking to be approved.\n\n`auth_result` and `decline_reason` are deliberately absent. The engine\ndecides them, the row is written carrying the decision, and raw capture is\nappend-only \u2014 so there is no moment at which this charge existed as an\napproved transaction and was then downgraded.",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "amount_base": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Base"
          },
          "billing_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Country"
          },
          "card_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Id"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "counterparty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "entry_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Mode"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "mcc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mcc"
          },
          "merchant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Id"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "payee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payee Id"
          },
          "shipping_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Country"
          },
          "txn_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Txn Country"
          },
          "txn_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Txn Id"
          },
          "txn_lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Txn Lat"
          },
          "txn_lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Txn Lon"
          },
          "txn_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Txn Type"
          }
        },
        "required": [
          "amount"
        ],
        "title": "AuthorizationRequest",
        "type": "object"
      },
...
      "EventBatch": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventRecord"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "EventBatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinkBatch": {
        "additionalProperties": false,
        "properties": {
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkRecord"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Links",
            "type": "array"
          }
        },
        "required": [
          "links"
        ],
        "title": "LinkBatch",
        "type": "object"
      },
...
      "RescoreReport": {
        "additionalProperties": false,
        "description": "A full-population pass over one lane, against the rules as they stand now.\n\nA SIBLING OF CycleReport, not a widening of it, because the two answer\ndifferent questions and `CycleReport`'s fields would have to lie to carry\nthis one. A cycle reports `since` \u2014 the window it consumed \u2014 and the counts\nfrom the cluster and feature stages it ran on the way. A re-score has no\nwindow by construction (that is the entire point: it ignores watermarks) and\nruns neither of those stages. Published through CycleReport it would report\n`since: null, clusters: 0, feature_values: 0`, which reads as \"those stages\nran and found nothing\" rather than \"those stages were not part of this\", and\na payload whose zeros mean absence is the thing this project keeps refusing\nto ship.\n\n`totals` is the raw dict `run_lane` returns rather than a fixed set of named\ncounters, for the reason `scripts/run_cycle.py` gives for printing it the\nsame way: a writer that learns to count something new should show up here\nwithout an edit. The console renders `evaluations`, `decisions` and `alerts`\nand ignores the rest.",
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "lane": {
            "enum": [
              "inline_sync",
              "async"
            ],
            "title": "Lane",
            "type": "string"
          },
          "totals": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Totals",
            "type": "object"
          }
        },
        "required": [
          "lane",
          "as_of"
        ],
        "title": "RescoreReport",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Serves the frozen alert.v1 contract, plus its siblings queue.v1, executions.v1, kpis.v1, explanation.v1, dispositions.v1, simulation.v1, catalog.v1 and ingest.v1. Reads are open; the surfaces that leave a mark \u2014 writing a disposition, running the engine on demand, writing to the control plane, and letting a row into raw capture \u2014 require a bearer token, and everything that touches a rule or ingests requires the admin role. A published rule lands in shadow: it is scored and recorded, and it acts on nothing until it is promoted. POST /authorize is the one endpoint that can stop a charge: the engine decides before the row is committed, so a declined charge is never an approved transaction.",
    "title": "GlassBox API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/authorize": {
      "post": {
        "description": "Decide a charge, and write the row carrying the decision.\n\nThis is the only endpoint in the project that can stop something. The action\nprecedence \u00a77 describes has always chosen `challenge` or `block` correctly;\nwhat it has never had is a charge whose fate depended on the answer. Here it\ndoes: a preventive action means the row is committed as `declined`, and\nbecause raw capture is append-only there is no moment at which a blocked\ncharge existed as an approved transaction.\n\nSynchronous by design. The inline lane's budget is 50 ms p99 (\u00a72.1) and this\nis where that number stops being a design target and starts being something\na caller can measure \u2014 `latency_ms` rides on the response for exactly that\nreason, and it is honest about including the feature pass.",
        "operationId": "authorize_charge_authorize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationOutcome"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authorize Charge"
      }
    },
...
    "/cycle/rescore": {
      "post": {
        "description": "Run every active rule over every subject in the lane, ignoring watermarks.\n\nTHE VERB A NEWLY PROMOTED RULE NEEDS, and it did not exist. A rule authored\nin the console, simulated, published and promoted produced no alerts and\nmoved no KPI, and both halves of the reason were correct behaviour:\n`GET /cycle` is a status read that evaluates nothing, and `POST /cycle`\nconsumes what has ARRIVED \u2014 with every watermark at the frontier it reports\n\"nothing has arrived since the last cycle\" and stops, which is exactly right.\nThe cycle reacts to data, not to rules. A new rule is not new data.\n\nNOT A FLAG ON `POST /cycle`. \"Consume what arrived\" and \"re-score everything\nagainst the rules as they stand now\" are different questions, and this\nproject's precedent for two meanings is two endpoints \u2014 `/authorize` against\n`/simulate/transaction`, and the two ingest doors this module opens with. A\n`POST /cycle?full=true` would be one query parameter away from a scheduled\ntick silently re-scoring the population every thirty seconds.\n\nSynchronous, because it was measured rather than assumed: the async lane's\nfull population is 169 subjects in ~0.3s and the inline lane's is ~9,850 in\n~13s. A spinner covers both; a job id and a polling console would be\nmachinery bought for a wait that does not exist.\n\nThe watermark is advanced exactly as `scripts/run_cycle.py` advances it, and\nfor the same reason \u2014 a hand-run lane and a scheduled one leave the same\nstate, so the next background tick sees a population it has consumed rather\nthan one it re-scores from scratch.\n\nAS_OF IS THE LATER OF `reference_now()` AND THE FRONTIER, which\n`scripts/run_cycle.py` is not, and the difference is not cosmetic. `as_of`\nis what binds the point-in-time feature read, so a re-score at\n`reference_now()` on a database whose newest charge arrived AFTER it\nevaluates every subject as of an instant before that charge existed \u2014 the\ndemo burst, or anything else that came in through `/authorize`, is simply\ninvisible to the rules being re-applied. That is the exact failure this\nendpoint exists to fix, one layer down. The script gets away with it because\nit takes `--as-of` and is run by someone who knows the dataset; a button\ncannot ask.",
        "operationId": "rescore_population_cycle_rescore_post",
        "parameters": [
          {
            "in": "query",
            "name": "lane",
            "required": false,
            "schema": {
              "default": "async",
              "enum": [
                "inline_sync",
                "async"
              ],
              "title": "Lane",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RescoreReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rescore Population"
      }
    },
...
    "/ingest/events": {
      "post": {
        "description": "The behavioural log. Without it S-077 cannot fire, whatever the transfer\nlooks like: its first condition is a password reset N minutes before the\nmovement, and that fact lives nowhere else.",
        "operationId": "ingest_events_ingest_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Events"
      }
    },
    "/ingest/links": {
      "post": {
        "description": "Edges in the link layer, which is what the graph builder reads.\n\nThe only way to reach L-203. A ring is discovered from `entity_links`, so a\nhundred ingested transfers between four accounts produce no cluster and no\nnetwork subject to score \u2014 the `opened_on` edges are the pattern.",
        "operationId": "ingest_links_ingest_links_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Links"
      }
    },
    "/ingest/transactions": {
      "post": {
        "description": "Transactions that already settled, including declines.\n\nThe door for planting a card-testing burst: `merchant_decline_burst` counts\ndeclines, and an authorization request cannot describe one because the\nengine would be deciding it.",
        "operationId": "ingest_transactions_ingest_transactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Transactions"
      }
    },
...
    "/simulate/transaction": {
      "post": {
        "description": "Score a charge that never happened.\n\n**Admin only** (WEEK5-PLAN O4). This is the one simulation that FABRICATES\nAN EVENT, and an event is the thing this system's whole record is made of: a\npayload showing a score against a transaction id is one screenshot away from\nbeing read as something that occurred. Every other defence here is on the\npayload \u2014 `persisted: false`, the fabricated row echoed as inserted, four\nnamed limits \u2014 and the role is the defence that does not depend on anyone\nreading it.\n\nThe refusals are answers, not errors. An unknown card, a borrowed real\n`txn_id` and a fabricated `synthetic_label` all come back as a 422 listing\nevery problem at once, from a check that runs BEFORE the sandbox opens.",
        "operationId": "simulate_hypothetical_transaction_simulate_transaction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSimulation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Hypothetical Transaction"
      }
    }
  }
}